+++ src/main/java/com/jscd/app/lecture/course/dao/CourseDao.java
@@ -1,0 +1,13 @@
+package com.jscd.app.lecture.course.dao;
+
+import com.jscd.app.lecture.course.dto.CourseDto;
+
+import java.util.List;
+
+public interface CourseDao {
+    List<CourseDto> getCourseList() throws Exception;
+    int addCourse(CourseDto dto) throws Exception;
+    CourseDto readCourse(Integer courseCode) throws Exception;
+    int removeCourse(Integer courseCode) throws Exception;
+    int modifyCourse(CourseDto dto) throws Exception;
+}
