+++ src/main/resources/mapper/lecture/course/courseMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.jscd.app.lecture.admin.courseMapper">
+<mapper namespace="com.jscd.app.lecture.course.courseMapper">
 <!--    과정 전체 읽기-->
     <select id="selectAll" resultType="CourseDto">
         select courseCode, courseName, regDate, firstIdNo
