+++ src/main/resources/mybatis-config.xml
@@ -9,8 +9,8 @@
 		<typeAlias alias="InstructorMemberInfoDto" type="com.jscd.app.admin.dto.InstructorMemberInfoDto"/>
 		<typeAlias alias="StdManageDto" type="com.jscd.app.admin.dto.StdManageDto"/>
 		<typeAlias alias="StdMemberManageDto" type="com.jscd.app.admin.dto.StdMemberManageDto"/>
-		<typeAlias alias="CourseDto" type="com.jscd.app.lecture.admin.dto.CourseDto"/>
-		<typeAlias alias="RoundDto" type="com.jscd.app.lecture.admin.dto.RoundDto"/>
+		<typeAlias alias="CourseDto" type="com.jscd.app.lecture.course.dto.CourseDto"/>
+		<typeAlias alias="RoundDto" type="com.jscd.app.lecture.round.dto.RoundDto"/>
 		<typeAlias alias="bootCampDto" type="com.jscd.app.lecture.bootCamp.dto.BootCampDto"/>
 		<typeAlias alias="SeminarRegistDto" type="com.jscd.app.lecture.seminar.dto.SeminarRegistDto"/>
 		<typeAlias alias="SearchCondition" type="com.jscd.app.admin.domain.SearchCondition"/>
